standardise writer header format
